Style handler move StyleHandler.js, added method parent()
