+++ lib/graph3d/Graph3d.js
@@ -9,7 +9,7 @@ var Filter       = require('./Filter');
 var Slider       = require('./Slider');
 var StepNumber   = require('./StepNumber');
 var support      = require('./stylehandlers/Support');
-var StyleHandler = require('./StyleHandler');
+var StyleHandler = require('./stylehandlers/StyleHandler');
 
 /**
  *
@@ -1250,13 +1250,13 @@ Graph3d.prototype._redrawAxis = function() {
 
 /**
  * Calculate the color based on the given value.
- * @param {Number} H   Hue, a value be between 0 and 360
- * @param {Number} S   Saturation, a value between 0 and 1
- * @param {Number} V   Value, a value between 0 and 1
+ * @param {Number} h   Hue, a value be between 0 and 360
+ * @param {Number} s   Saturation, a value between 0 and 1
+ * @param {Number} v   Value, a value between 0 and 1
  */
 Graph3d.prototype._hsv2rgb = function(h, s, v) {
   var c  = v * s;
-  var hi = Math.floor(H/60);  // hi = 0,1,2,3,4,5
+  var hi = Math.floor(h/60);  // hi = 0,1,2,3,4,5
   var x  = c * (1 - Math.abs(((h/60) % 2) - 1));
 
   var r;
@@ -1653,20 +1653,20 @@ Graph3d.prototype._showTooltip = function (dataPoint) {
 
   // calculate sizes
   var contentWidth  = content.offsetWidth;
-  var contentHeight   = content.offsetHeight;
+  var contentHeight = content.offsetHeight;
   var lineHeight    = line.offsetHeight;
-  var dotWidth    = dot.offsetWidth;
+  var dotWidth      = dot.offsetWidth;
   var dotHeight     = dot.offsetHeight;
 
   var maxLeft = Math.max(dataPoint.screen.x - contentWidth / 2, 10);
-  left = Math.min(maxLeft, this.frame.clientWidth - 10 - contentWidth);
+  var left    = Math.min(maxLeft, this.frame.clientWidth - 10 - contentWidth);
 
-  line.style.left   = dataPoint.screen.x + 'px';
-  line.style.top    = (dataPoint.screen.y - lineHeight) + 'px';
-  content.style.left  = left + 'px';
-  content.style.top = (dataPoint.screen.y - lineHeight - contentHeight) + 'px';
-  dot.style.left    = (dataPoint.screen.x - dotWidth / 2) + 'px';
-  dot.style.top     = (dataPoint.screen.y - dotHeight / 2) + 'px';
+  line.style.left    = dataPoint.screen.x + 'px';
+  line.style.top     = (dataPoint.screen.y - lineHeight) + 'px';
+  content.style.left = left + 'px';
+  content.style.top  = (dataPoint.screen.y - lineHeight - contentHeight) + 'px';
+  dot.style.left     = (dataPoint.screen.x - dotWidth / 2) + 'px';
+  dot.style.top      = (dataPoint.screen.y - dotHeight / 2) + 'px';
 };
