+++ lib/graph3d/StepNumber.js
@@ -84,7 +84,7 @@ StepNumber.calculatePrettyStep = function (step) {
   // try three steps (multiple of 1, 2, or 5
   var step1 = Math.pow(10, Math.round(log10(step)));
   var step2 = 2 * Math.pow(10, Math.round(log10(step / 2)));
-  varstep5 = 5 * Math.pow(10, Math.round(log10(step / 5)));
+  var step5 = 5 * Math.pow(10, Math.round(log10(step / 5)));
 
   // choose the best step (closest to minimum step)
   var prettyStep = step1;
