+++ lib/graph3d/stylehandlers/BarSizeStyleHandler.js
@@ -34,8 +34,7 @@ function(graph3d, ctx, point, xWidth, yWidth) {
   xWidth = (graph3d.xBarWidth / 2) * minOffset / denom;
   yWidth = (graph3d.yBarWidth / 2) * minOffset / denom;
 
-  var parent = BarParentStyleHandler.prototype;
-  parent.redrawPoint.call(this, graph3d, ctx, point, xWidth, yWidth);
+  this.parent().redrawPoint.call(this, graph3d, ctx, point, xWidth, yWidth);
 };
