+++ lib/graph3d/stylehandlers/DotLineStyleHandler.js
@@ -29,8 +29,7 @@ function(graph3d, ctx, point, dotSize) {
   ctx.lineTo(point.screen.x, point.screen.y);
   ctx.stroke();
 
-  var parent = PointStyleHandler.prototype;
-  parent.redrawPoint.call(this, graph3d, ctx, point, dotSize);
+  this.parent().redrawPoint.call(this, graph3d, ctx, point, dotSize);
 };
