+++ lib/graph3d/stylehandlers/MatrixStyleHandler.js
@@ -90,10 +90,18 @@ MatrixStyleHandler.prototype.getDataPoints = function(graph3d, data) {
       var d = dataMatrix[x][y];
       if (!d) continue;
 
-      var dDown  = dataMatrix[x+1][y];
-      var dRight = dataMatrix[x  ][y+1];
-      d.pointRight = (x < dataMatrix.length-1)    ? dDown  : undefined;
-      d.pointTop   = (y < dataMatrix[x].length-1) ? dRight : undefined;
+      if (x < dataMatrix.length-1) {
+        d.pointRight = dataMatrix[x+1][y];
+      } else {
+        d.pointRight = undefined;
+      }
+
+      if (y < dataMatrix[x].length-1) {
+        d.pointTop = dataMatrix[x][y+1];
+      } else {
+        d.pointTop = undefined;
+      }
+
       d.pointCross =
         (x < dataMatrix.length-1 && y < dataMatrix[x].length-1) ?
           dataMatrix[x+1][y+1] :
