+++ lib/graph3d/stylehandlers/StyleHandler.js
@@ -131,17 +131,17 @@
 // [1] - TODO quote Sandi Metz.
 //
 ////////////////////////////////////////////////////////////////////////////////
-var support               = require ('./stylehandlers/Support');
-var BarStyleHandler       = require ('./stylehandlers/BarStyleHandler');
-var BarColorStyleHandler  = require ('./stylehandlers/BarColorStyleHandler');
-var BarSizeStyleHandler   = require ('./stylehandlers/BarSizeStyleHandler');
-var LineStyleHandler      = require ('./stylehandlers/LineStyleHandler');
-var DotStyleHandler       = require ('./stylehandlers/DotStyleHandler');
-var DotLineStyleHandler   = require ('./stylehandlers/DotLineStyleHandler');
-var DotSizeStyleHandler   = require ('./stylehandlers/DotSizeStyleHandler');
-var DotColorStyleHandler  = require ('./stylehandlers/DotColorStyleHandler');
-var GridStyleHandler      = require ('./stylehandlers/GridStyleHandler');
-var SurfaceStyleHandler   = require ('./stylehandlers/SurfaceStyleHandler');
+var support               = require ('./Support');
+var BarStyleHandler       = require ('./BarStyleHandler');
+var BarColorStyleHandler  = require ('./BarColorStyleHandler');
+var BarSizeStyleHandler   = require ('./BarSizeStyleHandler');
+var LineStyleHandler      = require ('./LineStyleHandler');
+var DotStyleHandler       = require ('./DotStyleHandler');
+var DotLineStyleHandler   = require ('./DotLineStyleHandler');
+var DotSizeStyleHandler   = require ('./DotSizeStyleHandler');
+var DotColorStyleHandler  = require ('./DotColorStyleHandler');
+var GridStyleHandler      = require ('./GridStyleHandler');
+var SurfaceStyleHandler   = require ('./SurfaceStyleHandler');
 
 
 /// enumerate the available styles
