+++ lib/graph3d/stylehandlers/Support.js
@@ -4,6 +4,7 @@ var StepNumber   = require('../StepNumber');
 function derive(child, parent) {
   child.prototype = Object.create(parent.prototype);
   child.prototype.constructor = child;
+  child.prototype.parent = function() { return parent.prototype; }
 }
